+++ src/main/java/net/kaczmarzyk/spring/data/jpa/domain/Join.java
@@ -58,9 +58,8 @@ public class Join<T> implements Specification<T>, Fake {
 			String[] pathToJoinOnSplittedByDot = pathToJoinSplittedByDot(pathToJoinOn);
 
 			String extractedAlias = pathToJoinOnSplittedByDot[0];
-			javax.persistence.criteria.Join<?, ?> evaluated = queryContext.getEvaluated(extractedAlias, root);
-
-			if (evaluated == null) {
+			
+			if (!queryContext.existsJoin(extractedAlias, root)) {
 				throw new IllegalArgumentException(
 						"Join definition with alias: '" + extractedAlias + "' not found! " +
 								"Make sure that join with the alias '" + extractedAlias +"' is defined before the join with path: '" + pathToJoinOn + "'"
@@ -70,7 +69,10 @@ public class Join<T> implements Specification<T>, Fake {
 			String extractedPathToJoin = pathToJoinOnSplittedByDot[1];
                 queryContext.putLazyVal(
                         alias,
-                        (r) -> evaluated.join(extractedPathToJoin, joinType)
+                        (r) -> {
+                        	javax.persistence.criteria.Join<?, ?> evaluated = queryContext.getEvaluated(extractedAlias, root);
+                        	return evaluated.join(extractedPathToJoin, joinType);
+                        }
                 );
 		}
 		return null;