+++ src/test/java/net/kaczmarzyk/spring/data/jpa/utils/SimpleSpecificationGenerator.java
@@ -16,19 +16,19 @@
 package net.kaczmarzyk.spring.data.jpa.utils;
 
 import net.kaczmarzyk.spring.data.jpa.domain.*;
-import net.kaczmarzyk.spring.data.jpa.web.WebRequestQueryContext;
+import net.kaczmarzyk.spring.data.jpa.web.DefaultQueryContext;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.OnTypeMismatch;
 import org.springframework.data.jpa.domain.Specification;
 
 import java.lang.reflect.Constructor;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static java.util.Objects.isNull;
-import static net.kaczmarzyk.spring.data.jpa.web.utils.NativeWebRequestBuilder.nativeWebRequest;
 
 public class SimpleSpecificationGenerator {
 
@@ -46,8 +46,9 @@ public class SimpleSpecificationGenerator {
 	public static final Integer SEED_2 = 2;
 
 	private static Converter defaultConverter = Converter.withTypeMismatchBehaviour(OnTypeMismatch.EMPTY_RESULT, null);
-	private static WebRequestQueryContext queryCtx = new WebRequestQueryContext(nativeWebRequest().build());
+	private static DefaultQueryContext queryCtx = new DefaultQueryContext();
 
+	private static String DEFAULT_DATE_FORMAT = "yyyy-MM-dd";
 	private static LocalDateTime GENERATOR_LOCAL_BASE_DATE_TIME = LocalDateTime.now();
 
 	/**
@@ -102,9 +103,13 @@ public class SimpleSpecificationGenerator {
 	private static String[] getArgs(Integer argsNumber, Integer seed) {
 		String[] args = new String[argsNumber];
 		for (int i = 0; i < argsNumber; i++) {
-			args[i] = GENERATOR_LOCAL_BASE_DATE_TIME.plusSeconds(seed).toString();
+			args[i] = formatLocalDateTime(GENERATOR_LOCAL_BASE_DATE_TIME.plusDays(seed), DEFAULT_DATE_FORMAT);
 		}
 		return args;
 	}
 
+	private static String formatLocalDateTime(LocalDateTime date, String dateFormatPattern) {
+		DateTimeFormatter formatter = DateTimeFormatter.ofPattern(dateFormatPattern);
+		return date.format(formatter);
+	}
 }